requirements.txt
autopep8==2.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
importlib-metadata==6.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python
protobuf==3.20.3
pycodestyle==2.10.0
PyJWT==2.6.0
PyMySQL==1.0.2
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.12
flask_sqlalchemy
flask_swagger_ui
mysql-connector
requests